Imported Upstream version 2019.04.73

+++ content/browser/service_worker/service_worker_database.cc
@@ -119,6 +119,10 @@ const int64_t kCurrentSchemaVersion = 2;
 
 namespace {
 
+// The data size is usually small, but the values are changed frequently. So,
+// set a low write buffer size to trigger compaction more often.
+constexpr size_t kWriteBufferSize = 512 * 1024;
+
 class ServiceWorkerEnv : public leveldb_env::ChromiumEnv {
  public:
   ServiceWorkerEnv() : ChromiumEnv("LevelDBEnv.ServiceWorker") {}
@@ -293,7 +297,7 @@ ServiceWorkerDatabase::ServiceWorkerDatabase(const base::FilePath& path)
       next_avail_registration_id_(0),
       next_avail_resource_id_(0),
       next_avail_version_id_(0),
-      state_(UNINITIALIZED) {
+      state_(DATABASE_STATE_UNINITIALIZED) {
   sequence_checker_.DetachFromSequence();
 }
 
@@ -397,6 +401,8 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::GetRegistrationsForOrigin(
     return status;
 
   std::string prefix = CreateRegistrationKeyPrefix(origin);
+
+  // Read all registrations.
   {
     std::unique_ptr<leveldb::Iterator> itr(
         db_->NewIterator(leveldb::ReadOptions()));
@@ -421,21 +427,34 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::GetRegistrationsForOrigin(
         break;
       }
       registrations->push_back(registration);
-
-      if (opt_resources_list) {
-        std::vector<ResourceRecord> resources;
-        status = ReadResourceRecords(registration, &resources);
-        if (status != STATUS_OK) {
-          registrations->clear();
-          opt_resources_list->clear();
-          break;
-        }
-        opt_resources_list->push_back(resources);
-      }
     }
   }
 
+  // Count reading all registrations as one "read operation" for UMA
+  // purposes.
   HandleReadResult(FROM_HERE, status);
+  if (status != STATUS_OK)
+    return status;
+
+  // Read the resources if requested. This must be done after the loop with
+  // leveldb::Iterator above, because it calls ReadResouceRecords() which
+  // deletes |db_| on failure, and iterators must be destroyed before the
+  // database.
+  if (opt_resources_list) {
+    for (const auto& registration : *registrations) {
+      std::vector<ResourceRecord> resources;
+      // NOTE: ReadResourceRecords already calls HandleReadResult() on its own,
+      // so to avoid double-counting the UMA, don't call it again after this.
+      status = ReadResourceRecords(registration, &resources);
+      if (status != STATUS_OK) {
+        registrations->clear();
+        opt_resources_list->clear();
+        break;
+      }
+      opt_resources_list->push_back(resources);
+    }
+  }
+
   return status;
 }
 
@@ -1006,6 +1025,27 @@ ServiceWorkerDatabase::DeleteUserDataByKeyPrefixes(
   return WriteBatch(&batch);
 }
 
+ServiceWorkerDatabase::Status ServiceWorkerDatabase::RewriteDB() {
+  DCHECK(sequence_checker_.CalledOnValidSequence());
+
+  Status status = LazyOpen(false);
+  if (IsNewOrNonexistentDatabase(status))
+    return STATUS_OK;
+  if (status != STATUS_OK)
+    return status;
+  if (IsDatabaseInMemory())
+    return STATUS_OK;
+
+  leveldb_env::Options options;
+  options.create_if_missing = true;
+  options.env = g_service_worker_env.Pointer();
+  options.write_buffer_size = kWriteBufferSize;
+
+  status = LevelDBStatusToServiceWorkerDBStatus(
+      leveldb_env::RewriteDB(options, path_.AsUTF8Unsafe(), &db_));
+  return status;
+}
+
 ServiceWorkerDatabase::Status
 ServiceWorkerDatabase::ReadUserDataForAllRegistrations(
     const std::string& user_data_name,
@@ -1152,8 +1192,12 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::GetPurgeableResourceIds(
 
 ServiceWorkerDatabase::Status ServiceWorkerDatabase::ClearPurgeableResourceIds(
     const std::set<int64_t>& ids) {
+  Status status = LazyOpen(false);
+  if (IsNewOrNonexistentDatabase(status))
+    return STATUS_OK;
+
   leveldb::WriteBatch batch;
-  Status status = DeleteResourceIdsInBatch(
+  status = DeleteResourceIdsInBatch(
       service_worker_internals::kPurgeableResIdKeyPrefix, ids, &batch);
   if (status != STATUS_OK)
     return status;
@@ -1163,8 +1207,12 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::ClearPurgeableResourceIds(
 ServiceWorkerDatabase::Status
 ServiceWorkerDatabase::PurgeUncommittedResourceIds(
     const std::set<int64_t>& ids) {
+  Status status = LazyOpen(false);
+  if (IsNewOrNonexistentDatabase(status))
+    return STATUS_OK;
+
   leveldb::WriteBatch batch;
-  Status status = DeleteResourceIdsInBatch(
+  status = DeleteResourceIdsInBatch(
       service_worker_internals::kUncommittedResIdKeyPrefix, ids, &batch);
   if (status != STATUS_OK)
     return status;
@@ -1237,7 +1285,7 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::LazyOpen(
   DCHECK(sequence_checker_.CalledOnValidSequence());
 
   // Do not try to open a database if we tried and failed once.
-  if (state_ == DISABLED)
+  if (state_ == DATABASE_STATE_DISABLED)
     return STATUS_ERROR_FAILED;
   if (IsOpen())
     return STATUS_OK;
@@ -1257,9 +1305,7 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::LazyOpen(
   } else {
     options.env = g_service_worker_env.Pointer();
   }
-  // The data size is usually small, but the values are changed frequently. So,
-  // set a low write buffer size to trigger compaction more often.
-  options.write_buffer_size = 512 * 1024;
+  options.write_buffer_size = kWriteBufferSize;
 
   Status status = LevelDBStatusToServiceWorkerDBStatus(
       leveldb_env::OpenDB(options, path_.AsUTF8Unsafe(), &db_));
@@ -1277,7 +1323,7 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::LazyOpen(
   switch (db_version) {
     case 0:
       // This database is new. It will be initialized when something is written.
-      DCHECK_EQ(UNINITIALIZED, state_);
+      DCHECK_EQ(DATABASE_STATE_UNINITIALIZED, state_);
       return STATUS_OK;
     case 1:
       // This database has an obsolete schema version. ServiceWorkerStorage
@@ -1287,7 +1333,7 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::LazyOpen(
       return status;
     case 2:
       DCHECK_EQ(db_version, service_worker_internals::kCurrentSchemaVersion);
-      state_ = INITIALIZED;
+      state_ = DATABASE_STATE_INITIALIZED;
       return STATUS_OK;
     default:
       // Other cases should be handled in ReadDatabaseVersion.
@@ -1300,7 +1346,7 @@ bool ServiceWorkerDatabase::IsNewOrNonexistentDatabase(
     ServiceWorkerDatabase::Status status) {
   if (status == STATUS_ERROR_NOT_FOUND)
     return true;
-  if (status == STATUS_OK && state_ == UNINITIALIZED)
+  if (status == STATUS_OK && state_ == DATABASE_STATE_UNINITIALIZED)
     return true;
   return false;
 }
@@ -1773,14 +1819,14 @@ ServiceWorkerDatabase::Status ServiceWorkerDatabase::ReadDatabaseVersion(
 ServiceWorkerDatabase::Status ServiceWorkerDatabase::WriteBatch(
     leveldb::WriteBatch* batch) {
   DCHECK(batch);
-  DCHECK_NE(DISABLED, state_);
+  DCHECK_NE(DATABASE_STATE_DISABLED, state_);
 
-  if (state_ == UNINITIALIZED) {
+  if (state_ == DATABASE_STATE_UNINITIALIZED) {
     // Write database default values.
     batch->Put(
         service_worker_internals::kDatabaseVersionKey,
         base::Int64ToString(service_worker_internals::kCurrentSchemaVersion));
-    state_ = INITIALIZED;
+    state_ = DATABASE_STATE_INITIALIZED;
   }
 
   Status status = LevelDBStatusToServiceWorkerDBStatus(
@@ -1833,7 +1879,7 @@ void ServiceWorkerDatabase::Disable(const base::Location& from_here,
                 << " with error: " << StatusToString(status);
     DLOG(ERROR) << "ServiceWorkerDatabase is disabled.";
   }
-  state_ = DISABLED;
+  state_ = DATABASE_STATE_DISABLED;
   db_.reset();
 }
